requirements.txt
PyYAML>=6.0
asciinema>=2.0.0
